refactor: move all current pages to v1 version
