+++ src/app/not-found.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import Error from "next/error";
+
+export default function NotFound() {
+  return (
+    <html lang="en">
+      <body>
+        <Error statusCode={404} />
+      </body>
+    </html>
+  );
+}
