+++ src/app/v1/layout.tsx
@@ -1,0 +1,30 @@
+import type { Metadata } from "next";
+import "../globals.css";
+import { ClerkProvider } from "@clerk/nextjs";
+import Header from "@/components/Header";
+import { Toaster } from "react-hot-toast";
+
+export const metadata: Metadata = {
+  title: "Prompt.me",
+  description: "Compare chatbot models",
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <ClerkProvider>
+      <html lang="en" className="h-full">
+        <body className="flex flex-col h-full">
+          <Header />
+          <div className="flex flex-col h-full flex-1 bg-slate-200">
+            {children}
+          </div>
+          <Toaster position="top-right" />
+        </body>
+      </html>
+    </ClerkProvider>
+  );
+}
