+++ src/app/v1/page.tsx
@@ -1,0 +1,5 @@
+import Home from "@/components/Home";
+
+export default function HomePage() {
+  return <Home />;
+}
